requirements.txt
httpx>=0.24.0
cloudevents>=1.9.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
